Improve tigrc, add cl abbr shorten hirn to hrn
tigrc is now configured to not only open the selected file but also go
to the line number that is selected in tig at the time of opening the
file.
Also, add an abbreviation for the colorize-output-lines shell script, in
order to have it handy at any time.
Change hirn and related fish functions to hrn and omit the i, since it
is just shorter, therefore should type faster and ultimately should be
less error prone, since less keys are touched. The results are to be
awaited for but I am looking forward to it.