Rename grep -Hirns functions
Names are now more readable. Use abbreviations to have shortcut versions
of them.  Also, change minimerk's and bucket's and journal's
abbreviations to not append one " at the beginning of their command but
a ', which allows for raw interpretation of otherwise to-be-escaped
characters, which is always a small pain. This way, little less pain
everytime I am using it.

Also, come up with even shorter abbreviations.