Add abreviation xx for xargs -o vim -p
it's slightly faster compared to xv.

Also, add a special shortcut for sorting in visual mode that has the
same key-mapping like the sort-in-paragraph command which works in the
other modes. However, sort-in-paragraph has some issues when used on
multiple selected lines, therefore overwrite it for visual mode to be
mapped to a plain :sort<CR>.

Also add the bta (build/test all) abbreviation for a one liner that
builds ivi-mobile sparta for android & swift an runs the CPP unit tests
as well as the Android tests.