Add command AutoformatToggle
Add command AutoformatToggle that toggles the formatoptions' 'a'-flag
and harness this AutoformatToggle with a key-mapping '1' in normal mode.
This way, one can easily decide whether or not to autoformat the current
paragragh when writing or not.

Also move the 'commands' section on top of the key bindings and below
the 'functions' section, so the commands can be used direclty by the
key-bindings.