Add ranger config to dotfiles
Also update link-files.sh to wrap every path and variablie usage into
quotation marks " ". This way, it is safe to use variables and pathnames
that contain spaces.  Also, remove the if-blocks in the sublime
text-related passages that distinguis whether the original folders are
already links or real folders and trigger the according 'rm' command. In
fact, 'rm -rf' can delete links as well, but one has to take care not to
have trailing slashes '/' at the end of the paths. If, then not the link
will be deleted by rm -rf, but the target it is pointing to. The way it
is done here is the safe way, I suppose.