Fix a bug with autoformatting
Fix a bug that makes the behavior of the command AutoFormatToggle
unreliable when used the first few times. Interestingly, the problem
does not apper to be solved still for Git's COMMIT_EDIT_MSG files, i.e.
for Git's commit messages the problem still persists. I suppose, when
used with Git's commit messages, vim loads an extra file and does extra
line length magic, which is nice, but transparent to me.