Introduce vimlike-bookmarks.fish
Introduces a fast command for setting bookmarks that let you cd to the
current folder fast. The command is inspired by vim's and ranger's
commands for bookmarks: m, marks, delmarks and finally the cd-ing
commands `<char> and '<char>. Bookmarks are persistent and stored as
abbreviations, that means, they expand after entered, so they show what
they do. Also source the vimlike-bookmarks.fish inside the config.fish
and link it with the script link-files.sh.