Remove abbreviations for cd-ing into directories
Quickly jumping to certain directories shall now be done with the
functions of the script vimlike-bookmarks.fish and the abbreviations
that will be created by these functions. It is supposed to be faster and
its usage is similar of the the bookmarks-system found in the text
editor vim and the file explorer ranger.