Do not mention dotfiles in abbr fconf
Going to the .config/fish folder is not necessary but does not have to
happen via dotfiles, but using dotfiles's folder path implies a
restriction under which path dotfiles can be stored. This is not
necessary, since .config/fish folder can be accessed directly. This
eradicates the need for having a restriction on the dotfiles path.