Usue regex string contains for uname -n checks
uname -n encodes the node name, which may contain the computer's name,
but also the name of the network. Thus, checking for string-equality
with the computer-name won't work, rather check whether the computer
name is contained in the output of `uname -n`. Apparently, fish shell
does not support wildcard-matching if clauses, thus use switch-case
statements.