Create folder that contains sublime text packages
This way, sublime text packages can be eventually linked into the
system, although the sublime text folder might not exist. This is useful
when setting up a system from ground up and sublime text may not
installed yet. Otherwise, the packages may not be found when installing
sublime text later.