Set environment variables VISUAL and EDITOR
After accidently setting ranger to use nano as the default editor, even
after uninstalling ranger and re-installing it via apt did not force
ranger to ask for an editor again.
Research on the net suggests to set the environment variables VISUAL and
EDITOR. This might also alleviate the issue of setting the editor inside
ranger upon new installation.