Change the keyboard shortcuts for pushing to master
There were already shortcuts for pushing to refs/for/master (F12) and to
refs/for/staging. Remove the refs/for/staging - related shortcut and
replace it with a command for pushing to master (F11)  Remove the
commands for pushing via the keys 0 and 9 as they are redundant with F11
and F12.