Workon bail out when right environment is set
If the given environment is set already, bail out instead of calling
conda deactivate and then conda activate <ENVIRONMENT>. This is supposed
to execute faster if already on the right environment. However, the call
to conda env list that happens prior to that step already consumes some
time. It is best not to call workon if already on the right
directory and project.