Make fish grep-hirns-xxx functions abbreviations
This works on Ubuntu but may cause issues on Mac's grep version, since
on the mac version, at least around 2016-2018, you needed to put a path
specifier, even if grep should run in the current path as the last
argument to the command. This is not necessary for the ubuntu grep
version.

Also abbreviate many --include arguments into one with argument with {}.
This is untested on Mac.

Also remove some old and unnecessary grep commands.
Add a new cfh abbreviation to grep config files.

Remove function fn2 which is essentially equal to the more commonly used
function fd.