Refactor $history[1] related fish abbreviatons
Also align the fish abbreviations xh and ho to do the same thing. Both
now write to the clipboard and to ~/.histout. The abbreviation go also
now writes its output to the clipboard. Remove vimh abbreviation since
it it is a duplication of the abbreviation vh. vimh was never used as
opposed to vh. Moreover, the abbreviation vimh contains a bug.