Change git rebase --interactive shortcut in tig
Change the shortcut from 1 to `. At the moment, shortcuts for pushing to
git are at the end of the number row (i.e. 9 and 0 and may likely be
amended by other push options put onto 8, ... Consequently, it may be
nice to conversely have git pull options set at 1,2,.... So make already
some space there. The ` key is easily reachbable, too :)