Handle words with dots in vim normal mode normal
Use autocmds to change the iskeyword list to accept dots '.' and minuses
'-' in insert mode in order to have extended autocompletion but in
remove them again when leaving insert mode in order to have normal
search and moving behaviour with movements like w or with word search.
Also enable S-Tab to have autocompletion working backwards.
The next step would be to have YouAutoCompleteMe in place, but for now,
I am quite happy.