Keep vim's cursor  on screen-line when scrolling
The default vim behavior is, to place the cursor on the screen top when
scrolling down and on the screen bottom when scrolling up, which I find
this feels very unnatural. Half-Page scrolling (ctrl+u / ctrl+d) keeps
the cursor on the same line. Overwrite the behavior when full page
scrolling in order to keep the cursor on the same line when scrolling.

See:
https://vim.fandom.com/wiki/Page_up/down_and_keep_cursor_position