Improve ToggleComment() in vimrc
Now see comment-strings without trailing space as a comment.

Still, issues could arise with shebang-lines like the following:
'#!/bin/bash'

Also, commenting, uncommenting jumps the cursor inside the line if the
line is populated. If the line is empty, the cursor jumps to next
non-empty character. I will see if this turns out to be an issue.