Simplify tmux status lines
With non-maximized windows, the status lines are too long to be
displayed, in which case tmux would disable both the left and the right
one. However, I never use the right status line so I set it to nothing.
It would be replaced by the default right status line if I deleted the
according lines.